refactor: 기본 화면 이름 변경 및 바텀 바 선언 위치 변경
