+++ app/src/main/java/com/example/swap_it/MainActivity.kt
@@ -3,12 +3,13 @@ package com.example.swap_it
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.example.swap_it.ui.menu.MenuScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            RootScreen().MenuScreen()
+            MenuScreen().MenuScreen()
         }
     }
 }
