--- app/src/main/java/com/example/swap_it/ui/constant/Const.kt
@@ -1,7 +1,0 @@
-package com.example.swap_it.ui.constant
-
-const val PRODUCT = "PRODUCT"
-const val SHOPPING = "SHOPPING"
-const val ADD = "ADD"
-const val CHAT = "CHAT"
-const val USER = "USER"
